backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dataclasses-json==0.6.7
distro==1.9.0
docx2txt==0.8
emoji==2.14.0
eval_type_backport==0.2.0
fastapi==0.114.2
filetype==1.2.0
frozenlist==1.4.1
greenlet==3.1.0
grpcio==1.66.1
grpcio-tools==1.66.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-qdrant==0.1.4
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.121
lxml==5.3.0
MarkupSafe==3.0.1
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
olefile==0.47
openai==1.45.1
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
portalocker==2.10.1
prisma==0.15.0
protobuf==5.28.2
psutil==6.0.0
pycparser==2.22
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
PyJWT==2.9.0
PyMySQL==1.1.1
pyodbc==5.1.0
pypdf==5.0.1
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.12
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.2
pywin32==306
PyYAML==6.0.2
qdrant-client==1.11.2
RapidFuzz==3.10.0
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.38.5
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
tomlkit==0.13.2
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.15.14
unstructured-client==0.26.1
urllib3==2.2.3
uvicorn==0.30.6
vercel_blob==0.3.0
watchdog==5.0.3
wrapt==1.16.0
XlsxWriter==3.2.0
yarl==1.11.1